ru-video-storage/requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
pycryptodome>=3.19.0
tqdm>=4.66.0
requests>=2.31.0

# Video processing
ffmpeg-python>=0.2.0

# Data serialization
msgpack>=1.0.7

# CLI
click>=8.1.0
colorama>=0.4.6

# HTTP clients for platform APIs
aiohttp>=3.9.0
httpx>=0.25.0

# Utils
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
